Revert "Revert "Completed reproduction of all requested figures""
This reverts commit abda6deb651b3d728eda007dcb2685b024919d95.
